requirements/development.txt
-r base.txt

flake8
freezegun
isort
pytest-django
pytest-mock
